+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Added tweak disabling item interaction on boiling Crucibles
 - Added tweak blacklisting entities from spreading Flux Phage
 - Added tweak disabling Thaumonomicon category zoom
+- Added tweak giving the Fortress Armor knockback resistance
 
 ---